[#2] Refactoring CodeDisplay to add line number
