+++ example/components/CodeDisplay/styles.css
@@ -1,0 +1,22 @@
+.line-number {
+  text-align: right;
+  display: inline-block;
+  color: rgba(0, 0, 0, 0.2);
+  width: 40px;
+  font-size: 12px;
+  margin-right: 12px;
+}
+
+[data-pseudo-content]::before,
+[data-pseudo-content--before]::before,
+[data-pseudo-content--after]::after {
+  content: attr(data-pseudo-content);
+}
+
+.code-table {
+  tab-size: 2;
+}
+
+.code-table td, th { /* table cells */
+  padding: 0px;
+}
